+++ cpp/jni/javet_exceptions.cpp
@@ -31,6 +31,8 @@ namespace Javet {
 			jclassJavetConverterException = (jclass)jniEnv->NewGlobalRef(jniEnv->FindClass("com/caoccao/javet/exceptions/JavetConverterException"));
 			jclassJavetExecutionException = (jclass)jniEnv->NewGlobalRef(jniEnv->FindClass("com/caoccao/javet/exceptions/JavetExecutionException"));
 			jmethodIDJavetExecutionExceptionConstructor = jniEnv->GetMethodID(jclassJavetExecutionException, "<init>", "(Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;IIIII)V");
+			jclassJavetTerminatedException = (jclass)jniEnv->NewGlobalRef(jniEnv->FindClass("com/caoccao/javet/exceptions/JavetTerminatedException"));
+			jmethodIDJavetTerminatedExceptionConstructor = jniEnv->GetMethodID(jclassJavetTerminatedException, "<init>", "(Z)V");
 			jclassJavetUnknownCompilationException = (jclass)jniEnv->NewGlobalRef(jniEnv->FindClass("com/caoccao/javet/exceptions/JavetUnknownCompilationException"));
 			jmethodIDJavetUnknownCompilationExceptionConstructor = jniEnv->GetMethodID(jclassJavetUnknownCompilationException, "<init>", "(Ljava/lang/String;)V");
 			jclassJavetUnknownExecutionException = (jclass)jniEnv->NewGlobalRef(jniEnv->FindClass("com/caoccao/javet/exceptions/JavetUnknownExecutionException"));
@@ -42,8 +44,8 @@ namespace Javet {
 			auto isolate = v8Context->GetIsolate();
 			v8::String::Value exceptionMessage(isolate, v8TryCatch.Exception());
 			jstring jStringExceptionMessage = jniEnv->NewString(*exceptionMessage, exceptionMessage.length());
-			auto compileErrorMessage = v8TryCatch.Message();
-			if (compileErrorMessage.IsEmpty()) {
+			auto v8LocalMessage = v8TryCatch.Message();
+			if (v8LocalMessage.IsEmpty()) {
 				jthrowable javetUnknownCompilationException = (jthrowable)jniEnv->NewObject(
 					jclassJavetUnknownCompilationException,
 					jmethodIDJavetUnknownCompilationExceptionConstructor,
@@ -51,9 +53,9 @@ namespace Javet {
 				jniEnv->Throw(javetUnknownCompilationException);
 			}
 			else {
-				v8::String::Utf8Value scriptResourceName(isolate, compileErrorMessage->GetScriptResourceName());
+				v8::String::Utf8Value scriptResourceName(isolate, v8LocalMessage->GetScriptResourceName());
 				jstring jStringScriptResourceName = jniEnv->NewStringUTF(*scriptResourceName);
-				v8::String::Value sourceLine(isolate, compileErrorMessage->GetSourceLine(v8Context).ToLocalChecked());
+				v8::String::Value sourceLine(isolate, v8LocalMessage->GetSourceLine(v8Context).ToLocalChecked());
 				jstring jStringSourceLine = jniEnv->NewString(*sourceLine, sourceLine.length());
 				jthrowable javetConverterException = (jthrowable)jniEnv->NewObject(
 					jclassJavetCompilationException,
@@ -61,11 +63,11 @@ namespace Javet {
 					jStringExceptionMessage,
 					jStringScriptResourceName,
 					jStringSourceLine,
-					compileErrorMessage->GetLineNumber(v8Context).FromMaybe(0),
-					compileErrorMessage->GetStartColumn(),
-					compileErrorMessage->GetEndColumn(),
-					compileErrorMessage->GetStartPosition(),
-					compileErrorMessage->GetEndPosition() );
+					v8LocalMessage->GetLineNumber(v8Context).FromMaybe(0),
+					v8LocalMessage->GetStartColumn(),
+					v8LocalMessage->GetEndColumn(),
+					v8LocalMessage->GetStartPosition(),
+					v8LocalMessage->GetEndPosition());
 				jniEnv->Throw(javetConverterException);
 				jniEnv->DeleteLocalRef(jStringSourceLine);
 				jniEnv->DeleteLocalRef(jStringScriptResourceName);
@@ -79,37 +81,46 @@ namespace Javet {
 
 		void ThrowJavetExecutionException(JNIEnv* jniEnv, const v8::Local<v8::Context>& v8Context, const v8::TryCatch& v8TryCatch) {
 			auto isolate = v8Context->GetIsolate();
-			v8::String::Value exceptionMessage(isolate, v8TryCatch.Exception());
-			jstring jStringExceptionMessage = jniEnv->NewString(*exceptionMessage, exceptionMessage.length());
-			auto compileErrorMessage = v8TryCatch.Message();
-			if (compileErrorMessage.IsEmpty()) {
-				jthrowable javetUnknownExecutionException = (jthrowable)jniEnv->NewObject(
-					jclassJavetUnknownExecutionException,
-					jmethodIDJavetUnknownExecutionExceptionConstructor,
-					jStringExceptionMessage);
-				jniEnv->Throw(javetUnknownExecutionException);
+			if (v8TryCatch.HasTerminated()) {
+				jthrowable javetTerminatedException = (jthrowable)jniEnv->NewObject(
+					jclassJavetTerminatedException,
+					jmethodIDJavetTerminatedExceptionConstructor,
+					v8TryCatch.CanContinue());
+				jniEnv->Throw(javetTerminatedException);
 			}
 			else {
-				v8::String::Utf8Value scriptResourceName(isolate, compileErrorMessage->GetScriptResourceName());
-				jstring jStringScriptResourceName = jniEnv->NewStringUTF(*scriptResourceName);
-				v8::String::Value sourceLine(isolate, compileErrorMessage->GetSourceLine(v8Context).ToLocalChecked());
-				jstring jStringSourceLine = jniEnv->NewString(*sourceLine, sourceLine.length());
-				jthrowable javetConverterException = (jthrowable)jniEnv->NewObject(
-					jclassJavetExecutionException,
-					jmethodIDJavetExecutionExceptionConstructor,
-					jStringExceptionMessage,
-					jStringScriptResourceName,
-					jStringSourceLine,
-					compileErrorMessage->GetLineNumber(v8Context).FromMaybe(0),
-					compileErrorMessage->GetStartColumn(),
-					compileErrorMessage->GetEndColumn(),
-					compileErrorMessage->GetStartPosition(),
-					compileErrorMessage->GetEndPosition() );
-				jniEnv->Throw(javetConverterException);
-				jniEnv->DeleteLocalRef(jStringSourceLine);
-				jniEnv->DeleteLocalRef(jStringScriptResourceName);
+				v8::String::Value exceptionMessage(isolate, v8TryCatch.Exception());
+				jstring jStringExceptionMessage = jniEnv->NewString(*exceptionMessage, exceptionMessage.length());
+				auto v8LocalMessage = v8TryCatch.Message();
+				if (v8LocalMessage.IsEmpty()) {
+					jthrowable javetUnknownExecutionException = (jthrowable)jniEnv->NewObject(
+						jclassJavetUnknownExecutionException,
+						jmethodIDJavetUnknownExecutionExceptionConstructor,
+						jStringExceptionMessage);
+					jniEnv->Throw(javetUnknownExecutionException);
+				}
+				else {
+					v8::String::Utf8Value scriptResourceName(isolate, v8LocalMessage->GetScriptResourceName());
+					jstring jStringScriptResourceName = jniEnv->NewStringUTF(*scriptResourceName);
+					v8::String::Value sourceLine(isolate, v8LocalMessage->GetSourceLine(v8Context).ToLocalChecked());
+					jstring jStringSourceLine = jniEnv->NewString(*sourceLine, sourceLine.length());
+					jthrowable javetConverterException = (jthrowable)jniEnv->NewObject(
+						jclassJavetExecutionException,
+						jmethodIDJavetExecutionExceptionConstructor,
+						jStringExceptionMessage,
+						jStringScriptResourceName,
+						jStringSourceLine,
+						v8LocalMessage->GetLineNumber(v8Context).FromMaybe(0),
+						v8LocalMessage->GetStartColumn(),
+						v8LocalMessage->GetEndColumn(),
+						v8LocalMessage->GetStartPosition(),
+						v8LocalMessage->GetEndPosition());
+					jniEnv->Throw(javetConverterException);
+					jniEnv->DeleteLocalRef(jStringSourceLine);
+					jniEnv->DeleteLocalRef(jStringScriptResourceName);
+				}
+				jniEnv->DeleteLocalRef(jStringExceptionMessage);
 			}
-			jniEnv->DeleteLocalRef(jStringExceptionMessage);
 		}
 
 		void ThrowJavetV8RuntimeLockConflictException(JNIEnv* jniEnv, const char* message) {